requirements.txt
streamlit
langchain
langchain_groq
langchain_community
langchain_huggingface
langgraph
pypdf
faiss-cpu
wikipedia
arxiv
python-dotenv